refactor: 패키지 및 클래스명 명확히 변경
