+++ src/main/java/roomescape/dto/auth/LoginRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto.login;
+package roomescape.dto.auth;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
